% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\docType{data}
\name{annotation}
\alias{annotation}
\title{Annotation Data Frame}
\format{
A dataframe with 5000 rows and 3 columns
}
\usage{
data(annotation)
}
\description{
Contains information about the genes from the counts matrix
For the exact preprocessing steps, check the \code{annotation.R} script in the \code{data-raw} directory.
}
\keyword{datasets}
