% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\docType{data}
\name{counts}
\alias{counts}
\title{Raw count Matrix}
\format{
A gene expression matrix with 5000 rows and 32 samples.
}
\usage{
data(counts)
}
\description{
Contains the raw counts of GSE99353 and GSE112575. Number of genes was reduced to the 5k most variable ones.
For the exact preprocessing steps, check the \code{counts.R} script in the \code{data-raw} directory.
}
\keyword{datasets}
