% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{enrichment_heatmap}
\alias{enrichment_heatmap}
\title{Generate a heatmap of GSEA results}
\usage{
enrichment_heatmap(
  gsea_df,
  comparisons_column,
  comparisons = NULL,
  ordering = NULL,
  top_n = 10,
  data_src = "KEGG",
  color = viridis::plasma(50, end = 0.75),
  legend_title = "-log10(FDR)",
  title = "",
  remove_string = NULL,
  ...
)
}
\arguments{
\item{gsea_df}{A dataframe which contains the results of \code{msdbulkseqtools::glattCameraPR}. For creation see \code{examples}}

\item{comparisons_column}{The name of the column which contains the different comparisons that you want to make.}

\item{comparisons}{A vector of names which correspond to the columns of the heatmap (i.e. the conditions you want to compare).}

\item{ordering}{By which condition should the p values of the displayed gene sets be ranked}

\item{top_n}{The number of gene sets you want to display.}

\item{data_src}{The data source for which you want to display the top enriched terms.}

\item{color}{A colormap for the heatmap}

\item{legend_title}{The title of the legend}

\item{title}{The title of the heatmap}

\item{remove_string}{A string that should be removed from all rownames of the heatmap.}

\item{...}{Arguments are directly passed to \code{ComplexHeatmap::Heatmap}}
}
\description{
This function can be used to compare GSEA results performed for different tissues, different timepoints or conditions.
It will generate a heatmap, where rows represent different gene sets and columns the different comparisons.
}
\examples{
\dontrun{
Also see vignette `Standard bulk RNA seq showcase`
gsrc <- c("go_component","KEGG")
gsea <- msdbulkseqtools::glattCameraPR(g_src = gsrc,
                                      efit_list = res_list,
                                      specie = "Hs",
                                      g_cor_list = tcor_list,
                                      gs_low = 5,
                                      gs_high = 500)

gsea_df <- lapply(gsea, function(Ontology){
  lapply(names(Ontology), function(Tissue){
    lapply(names(Ontology[[Tissue]]),function(Contrast){
      Ontology[[Tissue]][[Contrast]] \%>\% mutate(contrast = Contrast)
      }) \%>\% dplyr::bind_rows() \%>\% mutate(tissue = Tissue)
   }) \%>\% bind_rows()
 }) \%>\% bind_rows()

enrichment_heatmap(gsea_df = gsea_df, comparisons = c("cervical","lumbar"),
                  top_n = 20, data_src = "go_component", comparisons_column = "tissue",
                  title = "Tissue comparison")
}
}
