% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdata.R
\docType{data}
\name{mdata}
\alias{mdata}
\title{Example metadata}
\format{
A data frame 32 rows (corresponding to samples) and 7 columns.
}
\usage{
data(mdata)
}
\description{
Contains the metadata of experiment GSE99353 and GSE112575.
For the exact preprocessing steps, check the \code{mdata.R} script in the \code{data-raw} directory.
Each row in the mdata corresponds to one column in the \code{counts} dataset.
}
\keyword{datasets}
