% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Plot a pretty volcano plot}
\usage{
volcano_plot(
  de_res,
  title = NULL,
  subtitle = NULL,
  annotate_by = NULL,
  annotation_size = 5.2,
  padj_threshold = 0.05,
  logFC_threshold = 1,
  ymax = 10,
  xlim = c(-3, 3)
)
}
\arguments{
\item{de_res}{A dataframe created using the \code{msdbulkseqtools::toptable_for_all} command. Alternatively a data frame
containing the following columns: 'symbol' (corresponding to gene name), 'adj.P.Val', 'logFC'}

\item{title}{The title of the plot}

\item{subtitle}{The subtitle of the plot}

\item{annotate_by}{A vector with gene names which will be annotated in the volcano plot.
Gene names have to occur in the 'symbol' column of 'de_res'}

\item{annotation_size}{The font size of the annotation.}

\item{padj_threshold}{Threshold determining which genes should be counted as differentially expressed}

\item{logFC_threshold}{Threshold determining which genes should be counted as "strongly" differentially expressed}

\item{ymax}{Maximum of the y axis}

\item{xlim}{The x axis limits.}
}
\value{
A ggplot of a volcano plot.
}
\description{
This function plots a pretty volcano plot using the results of the output of \code{msdbulkseqtools::toptable_for_all}
}
\examples{
\dontrun{
Also see vignette `Standard bulk RNA seq showcase`
res <- limmaVoomFit(v, design, cmat)
res_table <- msdbulkseqtools::toptable_for_all(res)
volcano_plot(res_table, annotate_by=c("SOD1"))
}

}
