% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{create_expression_boxplot}
\alias{create_expression_boxplot}
\title{Create an expression boxplot to compare two conditions.}
\usage{
create_expression_boxplot(
  gene,
  SE,
  intgroup,
  test_comparison = NULL,
  DESEq_res = NULL,
  colors = NULL,
  x_lab = ggplot2::waiver(),
  y_lab = expression(log[2](CPM)),
  ymax = NULL,
  ymin = NULL
)
}
\arguments{
\item{gene}{The gene symbol of the gene for which the expression should be plotted. Must occur in the \code{symbol} column of \code{rowData()} of the \code{SummarizedExperiment}.}

\item{SE}{A SummarizedExperiment. Usually the same which was used to generate the DESEq2 object.}

\item{intgroup}{Counts are grouped by this variable. Has to occur in \code{colData()} of the \code{SummarizedExperiment}.}

\item{test_comparison}{Optionally, you can also specify if and how conditions should be compared. Should be a list which contains the comparisons that you are interested in.
See examples and \href{https://cran.r-project.org/web/packages/ggsignif/vignettes/intro.html}{here}.}

\item{DESEq_res}{By default, if you specify \code{test_comparison} a \code{wilcox.test} is performed.
This might be fine to get an intuition if expression differences between two conditions
are significantly different, but the proper way is to use the the p value that
was determined using \code{DESeq}. For that, you can simply specify an object that was created \code{DESeq2::results()} that contains the correct contrast/comparison.
Also has to contain column \code{symbol}.}

\item{colors}{Specify the colors of the boxplots. Number of colors must match number of factors in \code{intgroup}.}

\item{x_lab}{X axis label. Defaults to \code{intgroup}}

\item{y_lab}{Y axis label. Defaults to \code{log(CPM)}}

\item{ymax}{Max value of the Y axis. Sometimes useful to pick a different value than the default.}

\item{ymin}{Min value of the Y axis. Sometimes useful to pick a different value than the default.}
}
\value{
A ggplot of a boxplot
}
\description{
Given a properly formatted \code{SummarizedExperiment.}, this function can quickly plot boxplots to investigate the expression of specific genes between two conditions.
By default, CPM transformed counts are expected
}
\examples{
\dontrun{
See vignette `rna_seq_workflow`
}

}
