% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{enrichment_heatmap}
\alias{enrichment_heatmap}
\title{Generate a heatmap of GSEA results}
\usage{
enrichment_heatmap(
  gsea_df,
  comparisons_column,
  ordering = NULL,
  top_n = 10,
  color = viridis::plasma(50, end = 0.75),
  legend_title = "-log10(FDR)",
  title = "",
  remove_string = NULL,
  ...
)
}
\arguments{
\item{gsea_df}{A dataframe which contains the results of \code{clusterProfiler::GSEA}.}

\item{comparisons_column}{The name of the column which contains the different comparisons that you want to make.}

\item{ordering}{By which condition should the p values of the displayed gene sets be ranked}

\item{top_n}{The number of gene sets you want to display.}

\item{color}{A colormap for the heatmap}

\item{legend_title}{The title of the legend}

\item{title}{The title of the heatmap}

\item{remove_string}{A string that should be removed from all rownames of the heatmap.}

\item{...}{Arguments are directly passed to \code{ComplexHeatmap::Heatmap}}
}
\description{
This function can be used to compare GSEA results performed for different tissues, different timepoints or conditions.
It will generate a heatmap, where rows represent different gene sets and columns the different comparisons.
}
