% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_de_correlation}
\alias{plot_de_correlation}
\title{Plot correlation between two differential gene expression analyses}
\usage{
plot_de_correlation(
  result_df1,
  result_df2,
  title_result_1,
  title_result_2,
  col = "log2FoldChange",
  remove_outlier = T
)
}
\arguments{
\item{result_df1}{The first results dataframe. Usually created using the \code{DESeq2::results()} command.}

\item{result_df2}{The second results dataframe. Usually created using the \code{DESeq2::results()} command.}

\item{title_result_1}{Name of the first differential gene expression analysis. Will be the title on the x-axis.}

\item{title_result_2}{Name of the second differential gene expression analysis. Will be the title on the y-axis.}

\item{col}{Name of the column which will be used to perform the correlation analysis.}

\item{remove_outlier}{If true, only values between the 1st and 99th percentile are included in analysis.
For example \code{log2FoldChange} Has to occur in both \code{result_df1} and \code{result_df2}}
}
\description{
This function visualizes the correlation between two different differential gene expression analyses.
Might be used to compare two different tissues or two different timepoints and see how similar they behave.
}
